[code] update code for new g++ 4.7.2 compiler

+++ toulbar2/src/tb2store.hpp
@@ -114,7 +114,7 @@ public:
 	void store() {
 		index++;
 		if (index >= indexMax) realloc();
-		pointers[index] = (T *) base;
+		pointers[index] = (T *) (intptr_t) base;
 		base = index;
 	}
